refactor(mf2): Rename src/cst-parser/ -> src/cst/
