+++ packages/mf2-messageformat/src/cst/resource-option.test.ts
