+++ packages/mf2-messageformat/src/cst/util.ts
