+++ packages/mf2-messageformat/src/data-model/types.ts
@@ -1,4 +1,4 @@
-import type * as CST from '../cst-parser/cst-types.js';
+import type * as CST from '../cst/types.js';
 import { cst } from './from-cst.js';
 
 /** A node in a message data model */
